+++ src/main/java/com/k2cybersecurity/intcodeagent/logging/ThreadMapping.java
@@ -10,14 +10,14 @@ public class ThreadMapping {
 
 	private static ThreadMapping threadMapping;
 	
-	private Map<Pair<Long, Integer>, ConcurrentLinkedDeque<ThreadRequestData>> tempThreadRequestMap = new ConcurrentHashMap<>();
+	private Map<Pair<Long, Long>, ConcurrentLinkedDeque<ThreadRequestData>> tempThreadRequestMap = new ConcurrentHashMap<>();
 
 	private Map<Long, ConcurrentLinkedDeque<ThreadRequestData>> mappedThreadRequestMap = new ConcurrentHashMap<>();
 
 	/**
 	 * @return the tempThreadRequestMap
 	 */
-	public Map<Pair<Long, Integer>, ConcurrentLinkedDeque<ThreadRequestData>> getTempThreadRequestMap() {
+	public Map<Pair<Long, Long>, ConcurrentLinkedDeque<ThreadRequestData>> getTempThreadRequestMap() {
 		return tempThreadRequestMap;
 	}
